+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - The `--compiler` flag now additionally resolves the compiler path using the `PATH` environment variable (more easily enabling `elm-review --compiler lamdera` for instance).
 - The `--elm-format-path` flag now additionally resolves the path to `elm-format` using the `PATH` environment variable.
 - Fixed an issue where the initial rule created with `elm-review new-package` was always a module rule, even when it was requested to be a project rule. Thanks to [@mateusfpleite](https://github.com/mateusfpleite)!
+- Fixed an issue where files in source-directories under `elm-stuff/` would not be included in the analysis.
 - Plenty of behind the scenes improvement to the maintenance of the repository. Thanks to [@lishaduck](https://github.com/lishaduck) for all of those.
 
 ## [2.12.0] - 2024-06-14